+++ BilgeHotel/BilgeHotel.Entities/Concrete/Reservation.cs
@@ -13,7 +13,7 @@ namespace BilgeHotel.Entities.Concrete
         public int Id { get; set; }
         public int CustomerId { get; set; }
         public int EmployeeId { get; set; }
-        public decimal ExtraTotalPrice { get; set; } //her extradan ne kadar ve kaç adet harcanmışsa
+        public double ExtraTotalPrice { get; set; } //her extradan ne kadar ve kaç adet harcanmışsa
                                                      //Fiyatı Hesaplanıp toplam fiyatı buraya yazılacaktır.
         public virtual Customer Customer { get; set; }
         public virtual Employee Employee { get; set; }
